requirements.txt
schedule
feedparser
requests
beautifulsoup4
google-generativeai
notion-client
python-dotenv
